+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/aad/AADAuthenticationProperties.java
@@ -317,6 +317,9 @@ public class AADAuthenticationProperties implements InitializingBean {
         this.postLogoutRedirectUri = postLogoutRedirectUri;
     }
 
+    @Deprecated
+    @DeprecatedConfigurationProperty(
+        reason = "Remove data collection based on fixed endpoints, and support based on service endpoints.")
     public boolean isAllowTelemetry() {
         return allowTelemetry;
     }