+++ sdk/spring/azure-spring-cloud-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/servicebus/AzureServiceBusQueueAutoConfiguration.java
@@ -7,7 +7,6 @@ import com.azure.messaging.servicebus.ServiceBusProcessorClient;
 import com.azure.spring.cloud.context.core.config.AzureProperties;
 import com.azure.spring.cloud.context.core.impl.ServiceBusNamespaceManager;
 import com.azure.spring.cloud.context.core.impl.ServiceBusQueueManager;
-import com.azure.spring.cloud.telemetry.TelemetryCollector;
 import com.azure.spring.integration.servicebus.converter.ServiceBusMessageConverter;
 import com.azure.spring.integration.servicebus.factory.DefaultServiceBusQueueClientFactory;
 import com.azure.spring.integration.servicebus.factory.ServiceBusConnectionStringProvider;
@@ -26,8 +25,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.util.Assert;
 
-import javax.annotation.PostConstruct;
-
 import static com.azure.spring.cloud.autoconfigure.servicebus.ServiceBusUtils.getNamespace;
 
 /**
@@ -42,14 +39,6 @@ import static com.azure.spring.cloud.autoconfigure.servicebus.ServiceBusUtils.ge
 public class AzureServiceBusQueueAutoConfiguration {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AzureServiceBusQueueAutoConfiguration.class);
-
-    private static final String SERVICE_BUS_QUEUE = "ServiceBusQueue";
-    private static final String NAMESPACE = "Namespace";
-
-    @PostConstruct
-    public void collectTelemetry() {
-        TelemetryCollector.getInstance().addService(SERVICE_BUS_QUEUE);
-    }
 
     @Bean
     @ConditionalOnMissingBean
@@ -79,8 +68,6 @@ public class AzureServiceBusQueueAutoConfiguration {
         clientFactory.setNamespace(properties.getNamespace());
         clientFactory.setServiceBusNamespaceManager(namespaceManager);
         clientFactory.setServiceBusQueueManager(queueManager);
-
-        TelemetryCollector.getInstance().addProperty(SERVICE_BUS_QUEUE, NAMESPACE, getNamespace(connectionString));
 
         return clientFactory;
     }