+++ sdk/spring/azure-spring-cloud-context/src/main/java/com/azure/spring/cloud/autoconfigure/context/AzureContextAutoConfiguration.java
@@ -6,11 +6,9 @@ package com.azure.spring.cloud.autoconfigure.context;
 import com.azure.core.credential.TokenCredential;
 import com.azure.core.management.profile.AzureProfile;
 import com.azure.resourcemanager.AzureResourceManager;
-import com.azure.spring.cloud.autoconfigure.telemetry.SubscriptionSupplier;
 import com.azure.spring.cloud.context.core.api.CredentialsProvider;
 import com.azure.spring.cloud.context.core.config.AzureProperties;
 import com.azure.spring.identity.DefaultSpringCredentialBuilder;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -65,11 +63,4 @@ public class AzureContextAutoConfiguration {
                                                    .alternativePrefix(AzureProperties.PREFIX)
                                                    .build();
     }
-
-    @Bean
-    @ConditionalOnBean(AzureResourceManager.class)
-    public SubscriptionSupplier subscriptionSupplier(AzureResourceManager azureResourceManager) {
-        return azureResourceManager::subscriptionId;
-    }
-
 }