+++ sdk/spring/azure-spring-cloud-storage/src/main/java/com/azure/spring/cloud/autoconfigure/storage/AzureStorageQueueAutoConfiguration.java
@@ -11,7 +11,6 @@ import com.azure.spring.cloud.context.core.api.EnvironmentProvider;
 import com.azure.spring.cloud.context.core.impl.DefaultEnvironmentProvider;
 import com.azure.spring.cloud.context.core.impl.StorageAccountManager;
 import com.azure.spring.cloud.context.core.storage.StorageConnectionStringProvider;
-import com.azure.spring.cloud.telemetry.TelemetryCollector;
 import com.azure.spring.integration.storage.queue.StorageQueueOperation;
 import com.azure.spring.integration.storage.queue.StorageQueueTemplate;
 import com.azure.spring.integration.storage.queue.factory.DefaultStorageQueueClientFactory;
@@ -39,13 +38,6 @@ import java.util.Optional;
 @ConditionalOnProperty(name = "spring.cloud.azure.storage.account")
 @EnableConfigurationProperties(AzureStorageProperties.class)
 public class AzureStorageQueueAutoConfiguration {
-
-    private static final String STORAGE_QUEUE = "StorageQueue";
-
-    @PostConstruct
-    public void collectTelemetry() {
-        TelemetryCollector.getInstance().addService(STORAGE_QUEUE);
-    }
 
     @Bean
     @ConditionalOnMissingBean({ StorageQueueClientFactory.class, StorageAccountManager.class })