+++ sdk/spring/azure-spring-cloud-stream-binder-servicebus-queue/src/main/java/com/azure/spring/servicebus/stream/binder/config/ServiceBusQueueBinderConfiguration.java
@@ -8,10 +8,8 @@ import com.azure.spring.cloud.autoconfigure.context.AzureEnvironmentAutoConfigur
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusAutoConfiguration;
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusProperties;
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusQueueAutoConfiguration;
-import com.azure.spring.cloud.autoconfigure.servicebus.ServiceBusUtils;
 import com.azure.spring.cloud.context.core.impl.ServiceBusNamespaceManager;
 import com.azure.spring.cloud.context.core.impl.ServiceBusQueueManager;
-import com.azure.spring.cloud.telemetry.TelemetryCollector;
 import com.azure.spring.integration.servicebus.queue.ServiceBusQueueOperation;
 import com.azure.spring.servicebus.stream.binder.ServiceBusQueueMessageChannelBinder;
 import com.azure.spring.servicebus.stream.binder.properties.ServiceBusQueueExtendedBindingProperties;
@@ -24,8 +22,6 @@ import org.springframework.cloud.stream.binder.Binder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-
-import javax.annotation.PostConstruct;
 
 /**
  * @author Warren Zhu
@@ -40,14 +36,6 @@ import javax.annotation.PostConstruct;
 @EnableConfigurationProperties({ AzureServiceBusProperties.class, ServiceBusQueueExtendedBindingProperties.class })
 public class ServiceBusQueueBinderConfiguration {
 
-    private static final String SERVICE_BUS_QUEUE_BINDER = "ServiceBusQueueBinder";
-    private static final String NAMESPACE = "Namespace";
-
-    @PostConstruct
-    public void collectTelemetry() {
-        TelemetryCollector.getInstance().addService(SERVICE_BUS_QUEUE_BINDER);
-    }
-
     @Bean
     @ConditionalOnMissingBean
     public ServiceBusChannelProvisioner serviceBusChannelProvisioner(
@@ -58,9 +46,6 @@ public class ServiceBusQueueBinderConfiguration {
         if (serviceBusNamespaceManager != null && serviceBusQueueManager != null) {
             return new ServiceBusQueueChannelResourceManagerProvisioner(serviceBusNamespaceManager,
                 serviceBusQueueManager, serviceBusProperties.getNamespace());
-        } else {
-            final String namespace = ServiceBusUtils.getNamespace(serviceBusProperties.getConnectionString());
-            TelemetryCollector.getInstance().addProperty(SERVICE_BUS_QUEUE_BINDER, NAMESPACE, namespace);
         }
         return new ServiceBusChannelProvisioner();
     }